+++ src/Questline/Cli/CliAppBuilder.cs
@@ -1,5 +1,6 @@
 using Microsoft.Extensions.DependencyInjection;
 using Questline.Engine;
+using Questline.Framework.Persistence;
 
 namespace Questline.Cli;
 
@@ -13,6 +14,7 @@ public class CliAppBuilder
         _services.AddSingleton<CliApp>();
 
         _services.AddQuestlineEngine();
+        _services.AddMongoPersistence("mongodb://localhost:27017", "questline", "0.6.0");
         return this;
     }